requirements.txt
python-docx
pillow
tkinterdnd2
